\name{calcD}
\alias{calcD}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
  Calculate Partition Factor
}
\description{
  \code{calcD} is an internal auxiliary function that calculates the
  partition factor, \eqn{D}, based on the number of slave processes, \eqn{P}.
}
\usage{
calcD(P)
}
\arguments{
  \item{P}{
    a positive integer, the number of slave processes.
  }
}