\name{distributedKrigeProblem-class}
\docType{class}
\alias{distributedKrigeProblem-class}
\alias{distributedKrigeProblem}
\title{
  ReferenceClass for Distributed Components of the krigeProblem ReferenceClass
}
\description{
  \code{distributedKrigeProblem} contains the distributed components of
  the core vectors and matrices of the krigeProblem class, as well as
  copies of the functions for calculating mean vectors and covariance
  matrices, parameter values, and information about the pieces of the
  distributed objects contained on a given slave process. The only
  method associated with the ReferenceClass is a constructor.
}
\seealso{
  \code{krigeProblem}
  }
