\name{localRm}
\alias{localRm}
\title{
  Remove Objects on Slave Process
}
\description{
  \code{localRm} is an internal auxiliary function used by
  \code{remoteRm} to remove objects on a slave process.
}
\usage{
localRm(list)
}
\arguments{
  \item{list}{a character vector naming objects to be removed.}
  }